FAB is disabled until data is loaded and adapter is its own file
